Move all boot code to one module
